+++ Sources/Streamer/Parser/Readium/ReadiumWebPubParser.swift
@@ -130,12 +130,18 @@ public class ReadiumWebPubParser: PublicationParser, Loggable {
                             $0.setPDFDocumentServiceFactory(DefaultPDFDocumentService.makeFactory(factory: pdfFactory))
                             $0.setTableOfContentsServiceFactory(LCPDFTableOfContentsService.makeFactory())
                             $0.setPositionsServiceFactory(LCPDFPositionsService.makeFactory())
+                            $0.setContentServiceFactory(DefaultContentService.makeFactory(
+                                resourceContentIteratorFactories: [
+                                    PDFResourceContentIterator.Factory(),
+                                ]
+                            ))
+                            $0.setSearchServiceFactory(ContentSearchService.makeFactory())
                         }
 
                         // FIXME: WebPositionsService from Kotlin?
 
                         if manifest.readingOrder.allAreHTML {
-                            $0.setSearchServiceFactory(StringSearchService.makeFactory())
+                            $0.setSearchServiceFactory(ContentSearchService.makeFactory())
                             $0.setContentServiceFactory(DefaultContentService.makeFactory(
                                 resourceContentIteratorFactories: [
                                     HTMLResourceContentIterator.Factory(),