+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 All notable changes to this project will be documented in this file. Take a look at [the migration guide](docs/Migration%20Guide.md) to upgrade between two major versions.
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+## [3.9.0] - 2026-05-12
 
 ### Added
 
@@ -16,6 +18,9 @@ All notable changes to this project will be documented in this file. Take a look
 
 * Added support for SVG covers in `ResourceCoverService`. SVG images can now be used as publication covers and are rendered to bitmaps (contributed by [@grighakobian](https://github.com/readium/swift-toolkit/pull/751)).
 * `Publication` has a new experimental `coverData(accepting:)` API that returns the raw bytes and media type of the cover, useful for storing the original cover without re-encoding.
+* New `ContentSearchService` implementation of `SearchService` that uses the Content API to search through publication resources.
+* PDF publications now support text extraction and search through the [Content](docs/Guides/Content.md) and [Search](docs/Guides/Search.md) APIs.
+    * **Note:** TTS support is not yet ready because sentence segmentation needs further refinement.
 
 #### Navigator
 
@@ -1223,3 +1228,4 @@ progression. Now if no reading progression is set, the `effectiveReadingProgress
 [3.6.0]: https://github.com/readium/swift-toolkit/compare/3.5.0...3.6.0
 [3.7.0]: https://github.com/readium/swift-toolkit/compare/3.6.0...3.7.0
 [3.8.0]: https://github.com/readium/swift-toolkit/compare/3.7.0...3.8.0
+[3.9.0]: https://github.com/readium/swift-toolkit/compare/3.8.0...3.9.0